+++ drift_cli/cli.py
@@ -6,6 +6,8 @@ from typing import Optional
 import click
 
 from drift_cli.alias import alias
+from drift_cli.export import export
+
 from drift_cli.config import write_config, Config
 
 
@@ -18,12 +20,6 @@ from drift_cli.config import write_config, Config
     help="Path to config file. Default ${HOME}/drift-cli/config.toml",
 )
 @click.option(
-    "--timeout",
-    "-t",
-    type=int,
-    help="Timeout for requests in seconds. Default 5",
-)
-@click.option(
     "--parallel",
     "-p",
     type=int,
@@ -33,15 +29,11 @@ from drift_cli.config import write_config, Config
 def cli(
     ctx,
     config: Optional[Path] = None,
-    timeout: Optional[int] = None,
     parallel: Optional[int] = None,
 ):
     """CLI client for PANDA | Drift Platform"""
     if config is None:
         config = Path.home() / ".drift-cli" / "config.toml"
-
-    if timeout is None:
-        timeout = 5
 
     if parallel is None:
         parallel = 10
@@ -50,8 +42,8 @@ def cli(
         write_config(config, Config(aliases={}))
 
     ctx.obj["config_path"] = config
-    ctx.obj["timeout"] = timeout
     ctx.obj["parallel"] = parallel
 
 
 cli.add_command(alias, "alias")
+cli.add_command(export, "export")