+++ drift_cli/utils/helpers.py
@@ -3,12 +3,13 @@ import asyncio
 import signal
 import time
 from asyncio import Semaphore, Queue
+from concurrent.futures import Executor
 from datetime import datetime
 from pathlib import Path
-from typing import Tuple, List
+from typing import Tuple
 
 from click import Abort
-from reduct import EntryInfo, Bucket
+from drift_client import DriftClient
 from rich.progress import Progress
 
 from drift_cli.config import read_config, Alias
@@ -32,24 +33,29 @@ def get_alias(config_path: Path, name: str) -> Alias:
 def parse_path(path) -> Tuple[str, str]:
     """Parse path ALIAS/RESOURCE"""
     args = path.split("/")
-    if len(args) != 2:
+    if len(args) > 2:
         raise RuntimeError(
             f"Path {path} has wrong format. It must be 'ALIAS/BUCKET_NAME'"
         )
+
+    if len(args) == 1:
+        args.append("panda")
+
     return tuple(args)
 
 
-async def read_records_with_progress(
-    entry: EntryInfo,
-    bucket: Bucket,
+async def read_topic(
+    pool: Executor,
+    client: DriftClient,
+    topic: str,
     progress: Progress,
     sem: Semaphore,
     **kwargs,
 ):  # pylint: disable=too-many-locals
     """Read records from entry and show progress
     Args:
-        entry (EntryInfo): Entry to read records from
-        bucket (Bucket): Bucket to read records from
+        client: Drift client
+        topic: Topic name
         progress (Progress): Progress bar to show progress
         sem (Semaphore): Semaphore to limit parallelism
     Keyword Args:
@@ -60,94 +66,82 @@ async def read_records_with_progress(
         Record: Record from entry
     """
 
-    def _to_timestamp(date: str) -> int:
-        return int(
-            datetime.fromisoformat(date.replace("Z", "+00:00")).timestamp() * 1000_000
-        )
+    def _to_timestamp(date: str) -> float:
+        return datetime.fromisoformat(date.replace("Z", "+00:00")).timestamp()
 
-    start = _to_timestamp(kwargs["start"]) if kwargs["start"] else entry.oldest_record
-    stop = _to_timestamp(kwargs["stop"]) if kwargs["stop"] else entry.latest_record
-
-    include = {}
-    for item in kwargs["include"]:
-        if item:
-            key, value = item.split("=")
-            include[key] = value
-
-    exclude = {}
-    for item in kwargs["exclude"]:
-        if item:
-            key, value = item.split("=")
-            exclude[key] = value
+    start = _to_timestamp(kwargs["start"])
+    stop = _to_timestamp(kwargs["stop"])
 
     last_time = start
-    task = progress.add_task(f"Entry '{entry.name}' waiting", total=stop - start)
+    task = progress.add_task(f"Topic '{topic}' waiting", total=stop - start)
     async with sem:
         exported_size = 0
         count = 0
         stats = []
         speed = 0
 
+        loop = asyncio.get_running_loop()
+
         def stop_signal():
             signal_queue.put_nowait("stop")
 
-        asyncio.get_event_loop().add_signal_handler(signal.SIGINT, stop_signal)
-        asyncio.get_event_loop().add_signal_handler(signal.SIGTERM, stop_signal)
+        loop.add_signal_handler(signal.SIGINT, stop_signal)
+        loop.add_signal_handler(signal.SIGTERM, stop_signal)
 
-        async for record in bucket.query(
-            entry.name,
-            start=start,
-            stop=stop,
-            include=include,
-            exclude=exclude,
-            ttl=kwargs["timeout"] * sem._value,  # pylint: disable=protected-access
-        ):
+        packages = await loop.run_in_executor(
+            pool, client.get_package_names, topic, start, stop
+        )
+        for package in sorted(packages):
+            drift_pkg = await loop.run_in_executor(pool, client.get_item, package)
             if signal_queue.qsize() > 0:
                 # stop signal received
                 progress.update(
                     task,
-                    description=f"Entry '{entry.name}' "
-                    f"(copied {count} records ({pretty_size(exported_size)}), stopped",
+                    description=f"Topic '{topic}' "
+                    f"(copied {count} packages ({pretty_size(exported_size)}), stopped",
                     refresh=True,
                 )
                 return
 
-            exported_size += record.size
-            stats.append((record.size, time.time()))
+            pkg_size = len(drift_pkg.blob)
+            timestamp = float(drift_pkg.package_id) / 1000
+            exported_size += pkg_size
+            stats.append((pkg_size, time.time()))
             if len(stats) > 10:
                 stats.pop(0)
 
             if len(stats) > 1:
                 speed = sum(s[0] for s in stats) / (stats[-1][1] - stats[0][1])
 
-            yield record
+            yield drift_pkg, task
 
+            count += 1
             progress.update(
                 task,
-                description=f"Entry '{entry.name}' "
-                f"(copied {count} records ({pretty_size(exported_size)}), "
+                description=f"Topic '{topic}' "
+                f"(copied {count} packages ({pretty_size(exported_size)}), "
                 f"speed {pretty_size(speed)}/s)",
-                advance=record.timestamp - last_time,
+                advance=timestamp - last_time,
                 refresh=True,
             )
-            last_time = record.timestamp
-            count += 1
+
+            last_time = timestamp
 
         progress.update(task, total=1, completed=True)
 
 
-def filter_entries(entries: List[EntryInfo], names: List[str]) -> List[EntryInfo]:
-    """Filter entries by names"""
-    if not names or len(names) == 0:
-        return entries
-
-    if len(names) == 1 and names[0] == "":
-        return entries
-
-    def _filter(entry):
-        for name in names:
-            if name == entry.name:
-                return True
-        return False
-
-    return list(filter(_filter, entries))
+# def filter_entries(entries: List[EntryInfo], names: List[str]) -> List[EntryInfo]:
+#     """Filter entries by names"""
+#     if not names or len(names) == 0:
+#         return entries
+#
+#     if len(names) == 1 and names[0] == "":
+#         return entries
+#
+#     def _filter(entry):
+#         for name in names:
+#             if name == entry.name:
+#                 return True
+#         return False
+#
+#     return list(filter(_filter, entries))